moving server config to the cloud

+++ commands/member.js
@@ -11,26 +11,43 @@ module.exports = {
      * @param {String[]} args 
      */
 
-    async execute(message, client, args){
-        if(!message.member.roles.cache.has("988346073531707392")) return
-        let num = 0
-        let count = message.guild.members.cache.filter(user => !user.user.bot && !user.roles.cache.has("991027323174289418")).size
-        const c = await message.channel.send(`${num}/${count}`)
-        message.guild.members.cache.filter(user => !user.user.bot && !user.roles.cache.has("991027323174289418")).forEach(async member => {
-            await member.roles.add("991027323174289418")
-            num+=1
-            c.edit(`${num}/${count}`)
-        })
+    async execute(message, client, args, _, __, server){
+        let access = false
+        if(message.author.id === message.guild.ownerId) access = true
+        else{
+            for (role of server.adminRoles){
+                if(message.member.roles.cache.has(role)){
+                    access = true
+                }
+            }
+        }
 
-        num = 0
-        count = message.guild.members.cache.filter(user => user.user.bot && !user.roles.cache.has("988390823232176158")).size
-        message.guild.members.cache.filter(user => user.user.bot && !user.roles.cache.has("988390823232176158")).forEach(async member => {
-            await member.roles.add("988390823232176158")
-            num+=1
-            c.edit(`${num}/${count}`)
-        })
-        message.delete()
-        c.edit(`${num}/${count}, all roles added`)
-        deleteMessage(c, 2000)
+        if(!access) return
+
+        let num = 0
+        if(server.memberRole){
+            let count = message.guild.members.cache.filter(user => !user.user.bot && !user.roles.cache.has(server.memberRole)).size
+            const c = await message.channel.send(`${num}/${count}`)
+            message.guild.members.cache.filter(user => !user.user.bot && !user.roles.cache.has(server.memberRole)).forEach(async member => {
+                await member.roles.add(server.memberRole)
+                num+=1
+                c.edit(`${num}/${count}`)
+            })
+            deleteMessage(c, 2000)
+        }
+
+        if(server.botRole){
+            num = 0
+            let count = message.guild.members.cache.filter(user => user.user.bot && !user.roles.cache.has(server.botRole)).size
+            const c = await message.channel.send(`${num}/${count}`)
+            message.guild.members.cache.filter(user => user.user.bot && !user.roles.cache.has(server.botRole)).forEach(async member => {
+                await member.roles.add(server.botRole)
+                num+=1
+                c.edit(`${num}/${count}`)
+            })
+            message.delete()
+            c.edit(`${num}/${count}, all roles added`)
+            deleteMessage(c, 2000)
+        }
     }
 }
