+++ commands/say.js
@@ -11,13 +11,25 @@ module.exports = {
      * @param {String[]} args 
      */
 
-    execute(message, client, args){
-        if(args[0] === undefined){
-            message.channel.send("you have to add some text").then(msg => deleteMessage(msg, 5000))
-            return
+    execute(message, client, args, _, __, server){
+        let access = false
+        if(message.author.id === message.guild.ownerId) access = true
+        else{
+            for (role of server.adminRoles){
+                if(message.member.roles.cache.has(role)){
+                    access = true
+                }
+            }
         }
+
+
         let des = ""
-        if ((message.member.roles.cache.has("988541799738015774") && message.guildId === "988446971461247027") || (message.member.roles.cache.has("988346073531707392") && message.guildId === "983033385721151538")){
+        if (access){
+            if(args[0] === undefined){
+                message.channel.send("you have to add some text").then(msg => deleteMessage(msg, 5000))
+                return
+            }
+            
             if (message.stickers.first() != undefined){
                 des += `\n<${message.stickers.first().name}>`
             }
